students/KragelAlina/task_02/src/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
flask-cors==6.0.1
pydantic==2.9.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytest==8.3.3
requests==2.32.3
flasgger==0.9.7.1